+++ protected/humhub/components/Migration.php
@@ -641,7 +641,10 @@ class Migration extends \yii\db\Migration
      */
     protected function isInitialInstallation(): bool
     {
-        return (!Yii::$app->installationState->hasState(InstallationState::STATE_INSTALLED));
+        return Yii::$app->runtimeCache->getOrSet(__METHOD__, function() {
+            Yii::$app->settings->reload();
+            return (!Yii::$app->installationState->hasState(InstallationState::STATE_INSTALLED));
+        });
     }
 
     /**